clean up DDLStatementDispatcher
and move ALTER TABLE related code to separate class as it contains the most logic
